[SYCL][NFC] Factor out empty kernel creation boilerplate
